move state down
